Simplify migi-game views
