--- entrypoints/twitch.content/views/miniGames/ChestControls.svelte
@@ -1,17 +1,0 @@
-<MiniGamesPanel isRunning={gameService.isGameActive} isSendEnabled={gameService.isGamePhase} {toggle} {participate} {Icon} {name} {...props}>
-    {#snippet gameIndicators()}
-        {#if gameService.isRoundRunning}
-            <MiniGamesText>
-                (<MiniGamesTimer timeout={gameService.timeUntilMessage} />)
-            </MiniGamesText>
-        {/if}
-    {/snippet}
-</MiniGamesPanel>
-
-<script lang="ts">
-import { MiniGamesTimer, MiniGamesPanel, MiniGamesText } from './basicComponents';
-import { useChestMiniGame } from './composables';
-
-const props = $props();
-const { toggle, participate, gameService, Icon, name } = useChestMiniGame();
-</script>
