+++ entrypoints/twitch.content/views/miniGames/MiniGamesControls.svelte
@@ -1,0 +1,11 @@
+<div class="flex items-center gap-[16px]">
+    <MiniGamesControlsHitsquad />
+    <MiniGamesControlsLoot />
+    <MiniGamesControlsChest />
+</div>
+
+<script lang="ts">
+import MiniGamesControlsHitsquad from './MiniGamesControlsHitsquad.svelte';
+import MiniGamesControlsLoot from './MiniGamesControlsLoot.svelte';
+import MiniGamesControlsChest from './MiniGamesControlsChest.svelte';
+</script>
