+++ entrypoints/twitch.content/views/miniGames/MiniGamesControlsIndicators.svelte
