+++ entrypoints/twitch.content/views/miniGames/MiniGamesControlsItem.svelte
