--- entrypoints/twitch.content/views/miniGames/basicComponents/MiniGamesText.svelte
@@ -1,7 +1,0 @@
-<span class="font-mono font-bold text-gray-600 dark:text-[#71717a] text-[13px] leading-none">
-    {@render props.children()}
-</span>
-
-<script lang="ts">
-const props = $props();
-</script>
