--- entrypoints/twitch.content/views/miniGames/basicComponents/index.ts
@@ -1,3 +1,0 @@
-export { default as MiniGamesText } from './MiniGamesText.svelte';
-export { default as MiniGamesTimer } from './MiniGamesTimer.svelte';
-export { default as MiniGamesPanel } from './MiniGamesPanel.svelte';
