--- entrypoints/twitch.content/views/miniGames/compact/CompactMiniGamesControls.svelte
@@ -1,11 +1,0 @@
-<div class="flex items-center gap-[16px]">
-    <CompactMiniGamesControlsHitsquad />
-    <CompactMiniGamesControlsLoot />
-    <CompactMiniGamesControlsChest />
-</div>
-
-<script lang="ts">
-import CompactMiniGamesControlsHitsquad from './CompactMiniGamesControlsHitsquad.svelte';
-import CompactMiniGamesControlsLoot from './CompactMiniGamesControlsLoot.svelte';
-import CompactMiniGamesControlsChest from './CompactMiniGamesControlsChest.svelte';
-</script>
