--- entrypoints/twitch.content/views/miniGames/compact/index.ts
@@ -1,1 +1,0 @@
-export { default as CompactMiniGamesControls } from './CompactMiniGamesControls.svelte'
