--- entrypoints/twitch.content/views/miniGames/composables/index.ts
@@ -1,3 +1,0 @@
-export * from './useHitsquadMiniGame.svelte';
-export * from './useLootMiniGame.svelte';
-export * from './useChestMiniGame.svelte';
