--- entrypoints/twitch.content/views/miniGames/composables/useLootMiniGame.svelte.ts
@@ -1,23 +1,0 @@
-import { getContext } from 'svelte';
-import { Gem } from '@lucide/svelte';
-import { LootGameService } from '@twitch/modules/miniGames';
-
-export const useLootMiniGame = () => {
-    const gameService = getContext<LootGameService>('loot');
-
-    function participate() {
-        gameService.participate();
-    }
-
-    function toggle(isEnabled: boolean) {
-        isEnabled ? gameService.start() : gameService.stop();
-    }
-
-    return {
-        name: 'Loot',
-        Icon: Gem,
-        gameService,
-        participate,
-        toggle
-    };
-}
