+++ entrypoints/twitch.content/views/miniGames/index.ts
@@ -1,4 +1,1 @@
-export { default as HitsquadControls } from './HitsquadControls.svelte';
-export { default as LootControls } from './LootControls.svelte';
-export { default as ChestControls } from './ChestControls.svelte';
-export * from './compact';
+export { default as MiniGamesControls } from './MiniGamesControls.svelte';
